requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------

matplotlib>=3.2.2
numpy>=1.19.5
opencv-python>=4.5.5.64
Pillow>=8.4.0
PyYAML>=6.0
requests>=2.27.1
scipy>=1.4.1
tqdm>=4.64.0

# plotting ------------------------------------

pandas>=1.1.4
seaborn>=0.11.0

# hybrid_model--------------------------------

Image
partial
argparse
tensorboard